Se realizan los cambios necesarios para accountspayable

+++ src/main/java/mx/bidg/service/impl/PriceEstimationsServiceImpl.java
@@ -132,13 +132,11 @@ public class PriceEstimationsServiceImpl implements PriceEstimationsService {
         if (request.getRequestStatus().getIdRequestStatus().equals(CEstimationStatus.PENDIENTE)) {
             
             String folio = request.getFolio();
-            List<PeriodicsPayments> periodicPayments = periodicPaymentsDao.findByFolio(folio);
-            for (PeriodicsPayments payment : periodicPayments) {
-                if (payment.getPeriodicPaymentStatus().getIdPeriodicPaymentStatus().equals(CPeriodicPaymentsStatus.INACTIVO)) {
-                    if(!periodicPaymentsDao.delete(payment))
-                        throw new ValidationException("No se pudo eliminar el PeriodicPayment: " + payment);
+            PeriodicsPayments periodicPayment = periodicPaymentsDao.findByFolio(folio);
+                if (periodicPayment.getPeriodicPaymentStatus().getIdPeriodicPaymentStatus().equals(CPeriodicPaymentsStatus.INACTIVO)) {
+                    if(!periodicPaymentsDao.delete(periodicPayment))
+                        throw new ValidationException("No se pudo eliminar el PeriodicPayment: " + periodicPayment);
                 }
-            }
             
             List<AccountsPayable> accountsPayable = accountsPayableDao.findByFolio(folio);
             for(AccountsPayable account : accountsPayable) {
@@ -170,8 +168,13 @@ public class PriceEstimationsServiceImpl implements PriceEstimationsService {
 
     @Override
     public boolean delete(Integer idEstimation) {
-//        PriceEstimations estimation = pr
-        return priceEstimationsDao.delete(new PriceEstimations(idEstimation));
+        PriceEstimations estimation = priceEstimationsDao.findByIdFetchRequestStatus(idEstimation);
+        if(estimation.getIdEstimationStatus() == CEstimationStatus.PENDIENTE){
+            return priceEstimationsDao.delete(new PriceEstimations(idEstimation));
+        } else {
+            throw new ValidationException("La cotizacion no tiene estatus de Pendiente", "Solo pueden" +
+                    " eliminarse Cotizaciones Pendientes de aprobacion", HttpStatus.CONFLICT);
+        }
     }
 
     @Override
